start loggerd tests
